+++ src/main/java/DAO/MultipleChoiceUnorderedResponseQuestionDao.java
@@ -27,7 +27,7 @@ public class MultipleChoiceUnorderedResponseQuestionDao implements QuestionDao {
 
             ResultSet rs = statement.getGeneratedKeys();
             rs.next();
-            int question_id = rs.getInt(1);
+            long question_id = rs.getLong(1);
 
             HashSet<String> legalAnswers = q.getLegalAnswers();
             HashSet<String> choices = q.getChoices();
@@ -35,22 +35,22 @@ public class MultipleChoiceUnorderedResponseQuestionDao implements QuestionDao {
                 PreparedStatement statement1 = conn.prepareStatement
                         ("Insert into multiple_choice_unordered_answers(answer_text, question_id, is_correct) values (?, ?, ?);");
                 statement1.setString(1, s);
-                statement1.setInt(2, question_id);
+                statement1.setLong(2, question_id);
                 statement1.setBoolean(3, legalAnswers.contains(s));
                 statement1.execute();
             }
     }
 
     @Override
-    public List<Question> getQuestions(int quizId) throws SQLException {
+    public List<Question> getQuestions(long quizId) throws SQLException {
         List<Question> result = new ArrayList<>();
             PreparedStatement st = conn.prepareStatement("select * from multiple_choice_unordered_questions WHERE  quiz_id = ?;");
-            st.setInt(1, quizId);
+            st.setLong(1, quizId);
             ResultSet res = st.executeQuery();
 
             while (res.next()) {
                 String text = res.getString("question_text");
-                int question_id = res.getInt("id");
+                long question_id = res.getLong("id");
                 String legalStm = "select * from multiple_choice_unordered_answers  WHERE question_id = ? AND is_correct;";
                 String choicesStm = "select * from  multiple_choice_unordered_answers  WHERE question_id = ?;";
                 HashSet<String> legalAnswers = getAnswers(question_id, legalStm, conn);
@@ -60,20 +60,20 @@ public class MultipleChoiceUnorderedResponseQuestionDao implements QuestionDao {
             }
         return result;
     }
-    private void insertAnswers(String st, Connection conn, int question_id, HashSet<String> answers) throws SQLException {
+    private void insertAnswers(String st, Connection conn, long question_id, HashSet<String> answers) throws SQLException {
         for(String s : answers){
             PreparedStatement statement1 = conn.prepareStatement(st);
             statement1.setString(1, s);
-            statement1.setInt(2, question_id);
+            statement1.setLong(2, question_id);
             statement1.execute();
         }
     }
 
 
-    private HashSet<String> getAnswers(int question_id, String s, Connection conn) throws SQLException {
+    private HashSet<String> getAnswers(long question_id, String s, Connection conn) throws SQLException {
         HashSet<String> result = new HashSet<>();
         PreparedStatement st = conn.prepareStatement(s);
-        st.setInt(1, question_id);
+        st.setLong(1, question_id);
         ResultSet res = st.executeQuery();
         while(res.next()){
             result.add(res.getString("answer_text"));
